@@ -4,7 +4,7 @@ module UserWarnable
   extend ActiveSupport::Concern
 
   included do
-    enum warning_type: {
+    enum :warning_type, {
       warning: 1,
       record: 2,
       ban: 3,
