@@ -7,7 +7,7 @@ module ArtistsHelper
     if artist
       link_to(artist.name, artist_path(artist))
     else
-      link = link_to(name, new_artist_path(name: name))
+      link = link_to(name, new_artist_path(artist: { name: name }))
       notice = tag.span("*", class: "new-artist", title: "No artist with this name currently exists.")
       "#{link} #{notice}".html_safe
     end
