@@ -29,14 +29,14 @@ module PostVersionsHelper
 
     diff[:added_tags].each do |tag_name|
       classes = diff[:obsolete_added_tags].include?(tag_name) ? "obsolete" : ""
-      changes << tag.ins(link_to("+#{tag_name}", posts_path(tags: tag_name)), class: classes)
+      changes << tag.ins(link_to_wiki_or_new("+#{tag_name}", tag_name), class: classes)
     end
     diff[:removed_tags].each do |tag_name|
       classes = diff[:obsolete_removed_tags].include?(tag_name) ? "obsolete" : ""
-      changes << tag.del(link_to("-#{tag_name}", posts_path(tags: tag_name)), class: classes)
+      changes << tag.del(link_to_wiki_or_new("-#{tag_name}", tag_name), class: classes)
     end
     diff[:unchanged_tags].each do |tag_name|
-      changes << tag.span(link_to(tag_name, posts_path(tags: tag_name)))
+      changes << tag.span(link_to_wiki_or_new(tag_name))
     end
 
     tag.span(safe_join(changes, " "), class: "diff-list")
@@ -47,13 +47,13 @@ module PostVersionsHelper
     changes = []
 
     diff[:added_locked_tags].each do |tag_name|
-      changes << tag.ins(link_to("+#{tag_name}", posts_path(tags: tag_name)))
+      changes << tag.ins(link_to_wiki_or_new("+#{tag_name}", tag_name))
     end
     diff[:removed_locked_tags].each do |tag_name|
-      changes << tag.del(link_to("-#{tag_name}", posts_path(tags: tag_name)))
+      changes << tag.del(link_to_wiki_or_new("-#{tag_name}", tag_name))
     end
     diff[:unchanged_locked_tags].each do |tag_name|
-      changes << tag.span(link_to(tag_name, posts_path(tags: tag_name)))
+      changes << tag.span(link_to_wiki_or_new(tag_name))
     end
 
     tag.span(safe_join(changes, " "), class: "diff-list")
