@@ -10,14 +10,14 @@ module Danbooru
       def has_bit_flags(attributes, options = {})
         field = options[:field] || :bit_flags
 
+        define_singleton_method("flag_value_for") do |key|
+          index = attributes.index(key)
+          raise IndexError if index.nil?
+          1 << index
+        end
+
         attributes.each.with_index do |attribute, i|
           bit_flag = 1 << i
-
-          define_singleton_method("flag_value_for") do |key|
-            index = attributes.index(key)
-            raise IndexError if index.nil?
-            1 << index
-          end
 
           define_method(attribute) do
             send(field) & bit_flag > 0
