@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class DanbooruLogger
-  def self.log(exception, expected: false, **params)
+  def self.log(exception, expected: false)
     if expected
       Rails.logger.info("#{exception.class}: #{exception.message}")
     else
@@ -9,18 +9,14 @@ class DanbooruLogger
       Rails.logger.error("#{exception.class}: #{exception.message}\n#{backtrace}")
     end
 
-    if defined?(::NewRelic) && !expected
-      ::NewRelic::Agent.notice_error(exception, expected: expected, custom_params: params)
-    end
+    Datadog::Tracing.active_span&.set_error(exception) unless expected
   end
 
   def self.initialize(user)
-    add_attributes("user.id" => user.id, "user.name" => user.name)
+    add_attributes("user.id" => user.id) unless user.is_anonymous?
   end
 
   def self.add_attributes(**)
-    return unless defined?(::NewRelic)
-
-    ::NewRelic::Agent.add_custom_attributes(**)
+    Datadog::Tracing.active_span&.set_tags(**)
   end
 end
