@@ -4,10 +4,11 @@ module PostSets
   class Favorites < PostSets::Base
     attr_reader :page, :limit
 
-    def initialize(user, page = 1, limit = CurrentUser.per_page)
+    def initialize(user, page, limit:)
+      super()
       @user = user
       @page = page
-      @limit = limit || CurrentUser.per_page
+      @limit = limit
     end
 
     def public_tag_string
@@ -21,7 +22,7 @@ module PostSets
     def posts
       @post_count ||= ::Post.tag_match("fav:#{@user.name} status:any").count_only
       @posts ||= begin
-        favs = ::Favorite.for_user(@user.id).includes(:post).order(created_at: :desc).paginate(page, total_count: @post_count, limit: @limit)
+        favs = ::Favorite.for_user(@user.id).includes(:post).order(created_at: :desc).paginate_posts(page, total_count: @post_count, limit: @limit)
         new_opts = { pagination_mode: :numbered, records_per_page: favs.records_per_page, total_count: @post_count, current_page: current_page }
         ::Danbooru::Paginator::PaginatedArray.new(favs.map(&:post), new_opts)
       end
