@@ -9,6 +9,10 @@ class ApplicationRecord < ActiveRecord::Base
         extending(Danbooru::Paginator::ActiveRecordExtension).paginate(page, options)
       end
 
+      def paginate_posts(page, options = {})
+        extending(Danbooru::Paginator::ActiveRecordExtension).paginate_posts(page, options)
+      end
+
       def qualified_column_for(attr)
         "#{table_name}.#{column_for_attribute(attr).name}"
       end
@@ -210,11 +214,10 @@ class ApplicationRecord < ActiveRecord::Base
         connection.execute("SET STATEMENT_TIMEOUT = #{CurrentUser.user.try(:statement_timeout) || 3_000}") unless Rails.env == "test"
       end
 
-      def with_timeout(n, default_value = nil, new_relic_params = {})
+      def with_timeout(n, default_value = nil)
         connection.execute("SET STATEMENT_TIMEOUT = #{n}") unless Rails.env == "test"
         yield
-      rescue ::ActiveRecord::StatementInvalid => x
-        DanbooruLogger.log(x, expected: true)
+      rescue ::ActiveRecord::StatementInvalid
         return default_value
       ensure
         connection.execute("SET STATEMENT_TIMEOUT = #{CurrentUser.user.try(:statement_timeout) || 3_000}") unless Rails.env == "test"
