@@ -4,8 +4,8 @@ class TagAlias < TagRelationship
   has_many :tag_rel_undos, as: :tag_rel
 
   after_save :create_mod_action
-  validates :antecedent_name, uniqueness: { conditions: -> { duplicate_relevant } }
-  validate :absence_of_transitive_relation
+  validates :antecedent_name, uniqueness: { conditions: -> { duplicate_relevant } }, unless: :is_deleted?
+  validate :absence_of_transitive_relation, unless: :is_deleted?
 
   module ApprovalMethods
     def approve!(update_topic: true, approver: CurrentUser.user)
@@ -137,9 +137,7 @@ class TagAlias < TagRelationship
     Post.without_timeout do
       Post.where_ilike(:locked_tags, "*#{consequent_name}*").find_each(batch_size: 50) do |post|
         fixed_tags = TagAlias.to_aliased_query(post.locked_tags, overrides: {consequent_name => antecedent_name})
-        CurrentUser.scoped(creator, creator_ip_addr) do
-          post.update_column(:locked_tags, fixed_tags)
-        end
+        post.update_column(:locked_tags, fixed_tags)
       end
     end
   end
@@ -172,9 +170,7 @@ class TagAlias < TagRelationship
   def rename_artist_undo
     if consequent_tag.category == Tag.categories.artist
       if consequent_tag.artist.present? && antecedent_tag.artist.blank?
-        CurrentUser.scoped(creator, creator_ip_addr) do
-          consequent_tag.artist.update!(name: antecedent_name)
-        end
+        consequent_tag.artist.update!(name: antecedent_name)
       end
     end
   end
@@ -208,8 +204,6 @@ class TagAlias < TagRelationship
         forum_updater.update(failure_message(e), "FAILED") if update_topic
         update_columns(status: "error: #{e}")
       end
-
-      DanbooruLogger.log(e, tag_alias_id: id, antecedent_name: antecedent_name, consequent_name: consequent_name)
     end
   end
 
@@ -253,6 +247,7 @@ class TagAlias < TagRelationship
   end
 
   def ensure_category_consistency
+    return if consequent_tag.post_count > 10_000 # Don't change category of large established tags.
     return if consequent_tag.is_locked? # Prevent accidentally changing tag type if category locked.
     return if consequent_tag.category != Tag.categories.general # Don't change the already existing category of the target tag
     return if antecedent_tag.category == Tag.categories.general # Don't set the target tag to general
@@ -273,9 +268,7 @@ class TagAlias < TagRelationship
     Post.without_timeout do
       Post.where_ilike(:locked_tags, "*#{antecedent_name}*").find_each(batch_size: 50) do |post|
         fixed_tags = TagAlias.to_aliased_query(post.locked_tags)
-        CurrentUser.scoped(creator, creator_ip_addr) do
-          post.update_column(:locked_tags, fixed_tags)
-        end
+        post.update_column(:locked_tags, fixed_tags)
       end
     end
   end
@@ -295,9 +288,7 @@ class TagAlias < TagRelationship
   def rename_artist
     if antecedent_tag.category == Tag.categories.artist
       if antecedent_tag.artist.present? && consequent_tag.artist.blank?
-        CurrentUser.scoped(creator, creator_ip_addr) do
-          antecedent_tag.artist.update!(name: consequent_name)
-        end
+        antecedent_tag.artist.update!(name: consequent_name)
       end
     end
   end
