@@ -14,7 +14,7 @@ class Ticket < ApplicationRecord
   validates :reason, presence: true
   validates :reason, length: { minimum: 2, maximum: Danbooru.config.ticket_max_size }
   validates :response, length: { minimum: 2 }, on: :update
-  enum status: %i[pending partial approved].index_with(&:to_s)
+  enum :status, %i[pending partial approved].index_with(&:to_s)
   after_update :log_update
   after_update :create_dmail
   validate :validate_content_exists, on: :create
