@@ -70,7 +70,7 @@ class UploadWhitelist < ApplicationRecord
         return [x.allowed, x.reason]
       end
     end
-    [false, "#{url.domain} not in whitelist"]
+    [false, "#{url.host} not in whitelist"]
   end
 
   extend SearchMethods
