@@ -62,6 +62,7 @@ class User < ApplicationRecord
   validates :email, presence: { if: :enable_email_verification? }
   validates :email, uniqueness: { case_sensitive: false, if: :enable_email_verification? }
   validates :email, format: { with: /\A.+@[^ ,;@]+\.[^ ,;@]+\z/, if: :enable_email_verification? }
+  validates :email, length: { maximum: 100 }
   validate :validate_email_address_allowed, on: [:create, :update], if: ->(rec) { (rec.new_record? && rec.email.present?) || (rec.email.present? && rec.email_changed?) }
 
   validates :name, user_name: true, on: :create
@@ -75,7 +76,6 @@ class User < ApplicationRecord
   validate :validate_ip_addr_is_not_banned, :on => :create
   validate :validate_sock_puppets, :on => :create, :if => -> { Danbooru.config.enable_sock_puppet_validation? }
   before_validation :normalize_blacklisted_tags, if: ->(rec) { rec.blacklisted_tags_changed? }
-  before_validation :set_per_page
   before_validation :staff_cant_disable_dmail
   before_validation :blank_out_nonexistent_avatars
   validates :blacklisted_tags, length: { maximum: 150_000 }
@@ -323,14 +323,6 @@ class User < ApplicationRecord
       can_approve_posts?
     end
 
-    def set_per_page
-      if per_page.nil?
-        self.per_page = Danbooru.config.posts_per_page
-      end
-
-      return true
-    end
-
     def blank_out_nonexistent_avatars
       if avatar_id.present? && avatar.nil?
         self.avatar_id = nil
@@ -384,8 +376,9 @@ class User < ApplicationRecord
 
     def is_blacklisting_user?(user)
       return false if blacklisted_tags.blank?
-      blta = blacklisted_tags.split("\n").map{|x| x.downcase}
-      blta.include?("user:#{user.name.downcase}") || blta.include?("uploaderid:#{user.id}")
+      bltags = blacklisted_tags.split("\n").map(&:downcase)
+      strings = %W[user:#{user.name.downcase} user:!#{user.id} userid:#{user.id}]
+      strings.any? { |str| bltags.include?(str) }
     end
   end
 
@@ -662,11 +655,11 @@ class User < ApplicationRecord
 
     # extra attributes returned for /users/:id.json but not for /users.json.
     def full_attributes
-      [
-        :wiki_page_version_count, :pool_version_count,
-        :forum_post_count, :comment_count,
-        :flag_count, :favorite_count, :positive_feedback_count,
-        :neutral_feedback_count, :negative_feedback_count, :upload_limit
+      %i[
+        wiki_page_version_count pool_version_count
+        forum_post_count comment_count flag_count favorite_count
+        positive_feedback_count neutral_feedback_count negative_feedback_count
+        upload_limit profile_about
       ]
     end
   end
