@@ -33,9 +33,7 @@ ForumCategory.find_or_create_by!(name: "Tag Alias and Implication Suggestions") 
 end
 
 def api_request(path)
-  response = HTTParty.get("https://e621.net#{path}", {
-    headers: { "User-Agent" => "e621ng/seeding" },
-  })
+  response = Faraday.get("https://e621.net#{path}", nil, user_agent: "e621ng/seeding")
   JSON.parse(response.body)
 end
 
@@ -83,6 +81,13 @@ end
 unless Rails.env.test?
   CurrentUser.user = admin
   CurrentUser.ip_addr = "127.0.0.1"
-  import_posts
-  import_mascots
+  begin
+    import_posts
+    import_mascots
+  rescue StandardError => e
+    puts "--------"
+    puts "#{e.class}: #{e.message}"
+    puts "Failure during seeding, continuing on..."
+    puts "--------"
+  end
 end
