@@ -7,14 +7,6 @@ class UserNameChangeRequestsControllerTest < ActionDispatch::IntegrationTest
     setup do
       @user = create(:privileged_user)
       @admin = create(:admin_user)
-      as(@user) do
-        @change_request = UserNameChangeRequest.create!(
-          :user_id => @user.id,
-          :original_name => @user.name,
-          :desired_name => "abc",
-          :change_reason => "hello"
-        )
-      end
     end
 
     context "new action" do
@@ -22,16 +14,43 @@ class UserNameChangeRequestsControllerTest < ActionDispatch::IntegrationTest
         get_auth new_user_name_change_request_path, @user
         assert_response :success
       end
-    end
 
-    context "create action" do
-      should "work" do
-        post_auth user_name_change_requests_path, @user, params: { user_name_change_request: { desired_name: "zun" }}
+      should "render for a user with a currently invalid username" do
+        @user.update_columns(name: "12345")
+        get_auth new_user_name_change_request_path, @user
         assert_response :success
       end
     end
 
+    context "create action" do
+      should "work" do
+        post_auth user_name_change_requests_path, @user, params: { user_name_change_request: { desired_name: "zun" } }
+        change_request = UserNameChangeRequest.last
+        assert_redirected_to user_name_change_request_path(change_request)
+        assert_equal("zun", @user.reload.name)
+      end
+
+      should "work for a user with a currently invalid name" do
+        @user.update_columns(name: "12345")
+        post_auth user_name_change_requests_path, @user, params: { user_name_change_request: { desired_name: "zun" } }
+        change_request = UserNameChangeRequest.last
+        assert_redirected_to user_name_change_request_path(change_request)
+        assert_equal("zun", @user.reload.name)
+      end
+    end
+
     context "show action" do
+      setup do
+        as(@user) do
+          @change_request = UserNameChangeRequest.create!(
+            user_id: @user.id,
+            original_name: @user.name,
+            desired_name: "abc",
+            change_reason: "hello",
+          )
+        end
+      end
+
       should "render" do
         get_auth user_name_change_request_path(@change_request), @user
         assert_response :success
