@@ -20,7 +20,7 @@ module PostSets
       context "a favorite set for before the most recent post" do
         setup do
           id = ::Favorite.where(user_id: @user.id, post_id: @post_3.id).first.id
-          @set = PostSets::Favorites.new(@user, "b#{id}", 1)
+          @set = PostSets::Favorites.new(@user, "b#{id}", limit: 1)
         end
 
         context "a sequential paginator" do
@@ -39,7 +39,7 @@ module PostSets
       context "a favorite set for after the third most recent post" do
         setup do
           id = ::Favorite.where(user_id: @user.id, post_id: @post_2.id).first.id
-          @set = PostSets::Favorites.new(@user, "a#{id}", 1)
+          @set = PostSets::Favorites.new(@user, "a#{id}", limit: 1)
         end
 
         context "a sequential paginator" do
@@ -52,7 +52,7 @@ module PostSets
       context "a favorite set for before the second most recent post" do
         setup do
           id = ::Favorite.where(user_id: @user.id, post_id: @post_1.id).first.id
-          @set = PostSets::Favorites.new(@user, "b#{id}", 1)
+          @set = PostSets::Favorites.new(@user, "b#{id}", limit: 1)
         end
 
         context "a sequential paginator" do
@@ -65,7 +65,7 @@ module PostSets
       context "a favorite set for after the second most recent post" do
         setup do
           id = ::Favorite.where(user_id: @user.id, post_id: @post_1.id).first.id
-          @set = PostSets::Favorites.new(@user, "a#{id}", 1)
+          @set = PostSets::Favorites.new(@user, "a#{id}", limit: 1)
         end
 
         context "a sequential paginator" do
@@ -77,7 +77,7 @@ module PostSets
 
       context "a favorite set for page 2" do
         setup do
-          @set = PostSets::Favorites.new(@user, 2, 1)
+          @set = PostSets::Favorites.new(@user, 2, limit: 1)
         end
 
         context "a numbered paginator" do
@@ -89,7 +89,7 @@ module PostSets
 
       context "a favorite set with no page specified" do
         setup do
-          @set = PostSets::Favorites.new(@user, nil, 1)
+          @set = PostSets::Favorites.new(@user, nil, limit: 1)
         end
 
         should "return the most recent element" do
